Moved all utility methods to utils module
